Added support for uploading multipart/related data
ObjectiveResource can now be confiured to upload multipart/related data.  These changes work hand-in-hand with updates I've made to the linked ObjectiveSupport sub-project, which now supports serialization of binary data (such as images).  These changes are a work in progress, as support still needs to be added for JSON serialzation and deserialzation in both XML and JSON.